supabase-metadata-events-update/requirements.txt
requests==2.31.0
python-dotenv>=0.19.0
supabase
aiohttp
web3